chg: dev: refactor credit and transfer modals in separate components
Furthers the codebase cleanup started in #249
