+++ src/assets/main.scss
@@ -623,7 +623,6 @@ rect {
   color: rgba(0, 0, 0, 0.4);
 }
 .custom-montant-input {
-  width: 89%;
 
   .frame3-sub-title {
     font-style: normal;
@@ -658,7 +657,7 @@ rect {
 
     resize: none;
     width: 100%;
-    height: 20vh;
+    height: 10vh;
   }
 
   textarea:focus-visible,
