+++ src/components/SendAskMoney.vue
@@ -6,9 +6,7 @@
           <button
             :disabled="!hasActiveMoneyAccount"
             @click="
-              ;(showModalFrame1 = true),
-                resetSendMoney(),
-                searchRecipients(),
+              ;(showTransferModal = true),
                 $store.commit('setModalState', true)
             "
             class="button custom-button is-payer has-text-weight-medium is-rounded action"
@@ -39,9 +37,7 @@
             :disabled="!hasActiveMoneyAccount"
             class="button custom-button is-recharger has-text-weight-medium is-rounded action"
             @click="
-              ;(showModalFrameCreditMoney1 = true),
-                setFocus(),
-                resetCredit(),
+              ;(showCreditModal = true),
                 $store.commit('setModalState', true)
             "
           >
@@ -56,365 +52,25 @@
       </div>
     </div>
 
-    <div class="modal is-active" v-if="showModalFrame1">
-      <div class="modal-background"></div>
-      <div class="modal-card">
-        <header class="modal-card-head">
-          <span class="is-flex is-justify-content-flex-start is-flex-shrink-0">
-            <a
-              class="mr-3 mt-1"
-              @click="
-                ;(showModalFrame1 = false),
-                  $store.commit('setModalState', false)
-              "
-            >
-              <img
-                class="cross-shape"
-                src="../assets/media/Arrow-Shape.png"
-                alt="arrow_shape"
-              />
-            </a>
-          </span>
-          <p class="modal-card-title is-title-shrink">
-            <span> Envoyer de l'argent </span>
-          </p>
-          <button
-            class="delete"
-            aria-label="close"
-            @click="
-              ;(showModalFrame1 = false), $store.commit('setModalState', false)
-            "
-          ></button>
-        </header>
-        <div class="search-area">
-          <div
-            class="mt-4 is-flex is-justify-content-space-evenly is-align-items-center custom-search-bar"
-          >
-            <p class="control has-icons-left custom-search-bar">
-              <input
-                v-model="searchName"
-                v-on:input="
-                  searchName.length === 0 || searchName.length >= 3
-                    ? searchRecipients()
-                    : null
-                "
-                class="input"
-                type="text"
-                placeholder="adresse mail, téléphone"
-              />
-              <span class="icon is-small is-left">
-                <i class="fas fa-search"></i>
-              </span>
-            </p>
-          </div>
-          <div
-            class="is-flex is-justify-content-space-evenly is-align-items-center mt-3"
-          ></div>
-          <div class="container is-fluid custom-heavy-line-separator"></div>
-        </div>
-        <section class="modal-card-body custom-modal-card-body">
-          <div v-if="isLoading" class="loader-container">
-            <loading
-              v-model:active="isLoading"
-              :can-cancel="false"
-              :is-full-page="false"
-              :width="50"
-              :height="50"
-            />
-          </div>
-          <div
-            v-else-if="searchRecipientError"
-            class="notification is-light is-danger"
-          >
-            Une erreur inattendue est survenue lors de la recherche de
-            destinataires. Veuillez nous excuser pour la gêne occasionnée.
-          </div>
-          <div v-else>
-            <div
-              v-if="ownCurrenciesPartners.length !== 0"
-              class="custom-card is-flex-direction-column is-align-items-center is-justify-content-space-between"
-            >
-              <div v-if="ownCurrenciesPartners">
-                <div
-                  class="pb-3 pt-3 is-flex flex-test is-flex-direction-row"
-                  v-for="partner in ownCurrenciesPartners"
-                  :key="partner"
-                >
-                  <div
-                    class="p-1 is-clickable is-align-items-center is-flex is-justify-content-flex-end favorit-icon-wrapper"
-                    :class="[partner.is_favorite ? 'is-active' : '']"
-                    @click="toggleFavorite(partner)"
-                  >
-                    <span>
-                      <i
-                        class="far fa-star"
-                        :class="[partner.is_favorite ? 'fas fa-star' : '']"
-                      ></i>
-                    </span>
-                  </div>
-                  <div
-                    class="p-2 is-clickable is-flex card-recipient-wrapper"
-                    @click="handleClickRecipient(partner)"
-                  >
-                    <span class="custom-card-destinataire">
-                      {{ partner.name }}
-                      {{
-                        partner.markBackend ? `(via ${partner.backendId})` : ""
-                      }}
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div
-              v-else
-              class="is-flex is-align-items-center is-justify-content-center"
-            >
-              Aucun destinataire de paiement a afficher
-            </div>
-          </div>
-        </section>
-        <footer class="modal-card-foot is-justify-content-flex-end">
-          <!--  <button class="button is-success">Save changes</button>
-        <button class="button">Cancel</button> -->
-        </footer>
-      </div>
-    </div>
-    <div class="modal is-active" v-if="showModalFrame2">
-      <div class="modal-background"></div>
-      <div class="modal-card">
-        <header class="modal-card-head">
-          <span class="is-flex is-flex-shrink-0">
-            <a
-              class="mr-3 mt-1"
-              @click=";(showModalFrame2 = false), (showModalFrame1 = true)"
-            >
-              <img
-                class="cross-shape"
-                src="../assets/media/Arrow-Shape.png"
-                alt="arrow_shape"
-              />
-            </a>
-          </span>
-          <p class="modal-card-title is-title-shrink">
-            <span> Envoyer à {{ recipientName }} </span>
-          </p>
-          <button
-            class="delete"
-            aria-label="close"
-            @click="
-              ;(showModalFrame2 = false), $store.commit('setModalState', false)
-            "
-          ></button>
-        </header>
-        <section class="modal-card-body">
-          <div>
-            <div
-              class="is-flex is-flex-direction-column is-justify-content-space-evenly is-align-items-center mt-3"
-            >
-              <div class="is-flex is-flex-direction-column custom-search-bar">
-                <h1 class="frame3-title mt-3 mb-3">Destinataire</h1>
-                <div
-                  class="is-flex is-justify-content-space-between is-align-items-center custom-user-select p-1"
-                >
-                  <div class="is-flex is-align-items-center">
-                    <figure class="image is-32x32">
-                      <img
-                        class="is-rounded"
-                        src="https://bulma.io/images/placeholders/128x128.png"
-                      />
-                    </figure>
-                    <h4 class="ml-1">{{ recipientName }}</h4>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div
-              class="is-flex is-flex-direction-column is-justify-content-space-evenly is-align-items-center mt-3"
-            >
-              <div
-                class="is-flex is-flex-direction-column custom-montant-input"
-              >
-                <h2 class="frame3-sub-title mt-3 mb-3">Montant</h2>
-                <div class="is-flex">
-                  <input
-                    v-model.number="amount"
-                    ref="amountSend"
-                    type="number"
-                    min="0"
-                    class="input is-custom"
-                    placeholder="ex: 50"
-                    :class="{
-                      'is-danger':
-                        errors.transfer.balance || errors.transfer.amount,
-                    }"
-                  />
-                  <div class="amount-currency-symbol pl-2">
-                    {{ recipientCurrencySymbol }}
-                  </div>
-                </div>
-                <div
-                  class="notification is-danger is-light"
-                  v-if="errors.transfer.balance"
-                >
-                  {{ errors.transfer.balance }}
-                </div>
-                <div
-                  class="notification is-danger is-light"
-                  v-if="errors.transfer.amount"
-                >
-                  {{ errors.transfer.amount }}
-                </div>
-                <textarea
-                  v-model="message"
-                  class="custom-textarea textarea mt-5"
-                  placeholder="Ajoutez un texte (optionnel)"
-                ></textarea>
-              </div>
-            </div>
-          </div>
-        </section>
-        <footer class="modal-card-foot is-justify-content-flex-end">
-          <button
-            class="button custom-button custom-button-send-receive-money is-rounded action"
-            @click="sendTransaction(), $store.commit('setModalState', false)"
-          >
-            Envoyer
-          </button>
-        </footer>
-      </div>
-    </div>
+    <money-transfer-modal
+      v-if="showTransferModal"
+      @close="
+        (showTransferModal = false),
+        $store.commit('setModalState', false)
+      "
+    />
+
+    <money-credit-modal
+      v-if="showCreditModal"
+      @close="
+        (showCreditModal = false),
+        $store.commit('setModalState', false)
+      "
+    />
+
   </div>
   <div ref="isLoadinMoneyContainer"></div>
-  <div
-    class="modal is-active"
-    v-if="showModalFrameCreditMoney1 || globalBalCall"
-  >
-    <div class="modal-background"></div>
-    <div class="modal-card">
-      <header class="modal-card-head">
-        <p class="modal-card-title is-title-shrink">Créditer mon compte</p>
-        <button
-          class="delete"
-          aria-label="close"
-          @click="
-            ;(showModalFrameCreditMoney1 = false),
-              (showCreditRefreshNotification = false),
-              resetCredit(),
-              $store.commit('setModalState', false)
-          "
-        ></button>
-      </header>
-      <section class="modal-card-body">
-        <div v-if="myHyperLink.length === 0 && !showCreditRefreshNotification">
-          <div v-if="creditableMoneyAccounts.length > 1">
-            <h2 class="frame3-sub-title mt-3 mb-3">Compte à créditer</h2>
-            <div
-              v-for="account in creditableMoneyAccounts"
-              :class="[
-                selectedCreditAccount === account ? 'selected' : 'unselected',
-                'account-selector',
-              ]"
-              @click="setSelectedCreditAccount(account), setFocus()"
-            >
-              <BankAccountItem
-                :bal="account.bal"
-                :curr="account.curr"
-                :backend="account.backend"
-                :type="account.type"
-                :active="account.active"
-              >
-                <template v-slot:name>{{ account.name }}</template>
-              </BankAccountItem>
-            </div>
-          </div>
-          <div
-            v-show="
-              selectedCreditAccount || creditableMoneyAccounts.length === 1
-            "
-            class="amount custom-montant-input w-100"
-          >
-            <h2 class="frame3-sub-title mt-3 mb-3">Montant à créditer</h2>
-            <div class="is-flex">
-              <input
-                v-model.number="amountForCredit"
-                ref="amountcredit"
-                type="number"
-                min="0"
-                class="input is-custom"
-                placeholder="ex: 50"
-                :class="{ 'is-danger': errors.credit.amount }"
-              />
-              <span class="amount-currency-symbol pl-2">{{
-                this.selectedCreditAccount?.curr
-              }}</span>
-            </div>
-            <div
-              class="notification is-danger is-light"
-              v-if="errors.credit.amount"
-            >
-              {{ errors.credit.amount }}
-            </div>
-          </div>
-        </div>
-        <template v-if="myHyperLink.length > 1">
-          <div class="notification is-info">
-            <p class="mb-3">
-              Un bon de commande pour votre rechargement a été créé.
-            </p>
-            <p class="mb-3">
-              Pour compléter la demande de crédit, vous devez finaliser la
-              transaction en vous rendant dans votre espace personnel Odoo:
-            </p>
-          </div>
-        </template>
-        <template v-if="showCreditRefreshNotification">
-          <div class="notification is-info">
-            <p class="mb-3" v-if="selectedCreditAccount.backend === 'comchain'">
-              Une fois votre opération complétée dans votre espace personnel,
-              votre crédit sera en attente de validation par un administrateur.
-              Vous pourrez alors fermer cette fenêtre pour actualiser votre
-              solde.
-            </p>
-            <p class="mb-3" v-if="selectedCreditAccount.backend === 'cyclos'">
-              Une fois votre opération complétée dans votre espace personnel,
-              fermez cette fenêtre pour actualiser votre solde.
-            </p>
-          </div>
-        </template>
-      </section>
-      <footer class="modal-card-foot is-justify-content-flex-end">
-        <template
-          v-if="
-            myHyperLink.length === 0 &&
-            (selectedCreditAccount || creditableMoneyAccounts.length === 1) &&
-            !showCreditRefreshNotification
-          "
-        >
-          <button
-            class="button custom-button custom-button-send-receive-money is-rounded action"
-            @click="newLinkTab()"
-          >
-            Suivant
-          </button>
-        </template>
-        <template v-if="myHyperLink.length > 1">
-          <a
-            class="button custom-button has-text-weight-medium custom-inverted is-rounded action"
-            @click="navigateToCreditOrder"
-            >Compléter la transaction dans mon espace personnel</a
-          >
-        </template>
-        <template v-if="showCreditRefreshNotification">
-          <a
-            class="button custom-button has-text-weight-medium custom-inverted is-rounded action"
-            @click="closeAndRefresh"
-            >Fermer et rafraîchir</a
-          >
-        </template>
-      </footer>
-    </div>
-  </div>
+
 </template>
 
 <script lang="ts">
@@ -422,64 +78,28 @@
   import { Options, Vue } from "vue-class-component"
   import { mapGetters, mapState } from "vuex"
 
-  import BankAccountItem from "./BankAccountItem.vue"
-  import Loading from "vue-loading-overlay"
-  import "vue-loading-overlay/dist/vue-loading.css"
-
-  function returnFavoritesOnly(partners: any): any {
-    var ret = []
-    for (let el of partners) {
-      if (el.is_favorite == true) {
-        ret.push(el)
-      }
-    }
-    return ret
-  }
+  import MoneyTransferModal from "./MoneyTransferModal.vue"
+  import MoneyCreditModal from "./MoneyCreditModal.vue"
 
   @Options({
     name: "SendAskMoney",
     components: {
-      BankAccountItem,
-      Loading,
+      MoneyTransferModal,
+      MoneyCreditModal,
     },
     data() {
       return {
-        showModalFrame1: false,
-        showModalFrame2: false,
+        showTransferModal: false,
+        showCreditModal: false,
         showModalFrameAskMoney1: false,
         showModalFrameAskMoney2: false,
-        showModalFrameCreditMoney1: false,
-        showModalFrameCreditMoney2: false,
-        showModalFrameCreditMoney3: false,
-        showModalFrameCreditMoney4: false,
         warning: true,
         activeClass: 0,
         favoris: false,
-        searchName: "",
-        amount: 0,
         message: "",
-        partners: [],
-        recipientName: "",
-        recipientCurrencySymbol: "",
-        displayFavoritesOnly: false,
         amountAsked: 0,
         linkGenerated: false,
         history: [],
-        amountForCredit: 0,
-        urlForHyperlink: "",
-        selectedCreditAccount: null,
-        showCreditRefreshNotification: false,
-        isLoading: false,
-        searchRecipientError: false,
-        errors: {
-          transfer: {
-            balance: false,
-            amount: false,
-          },
-          credit: {
-            amount: false,
-          },
-        },
       }
     },
 
@@ -487,94 +107,15 @@
       myLink(): string {
         return this.$store.state.lokapi.paymentUrl.order_url
       },
-      myHyperLink(): string {
-        return this.urlForHyperlink
-      },
       globalBalCall(): boolean {
         return this.$store.state.showCredit
       },
       hasActiveMoneyAccount(): boolean {
         return this.$store.getters.activeVirtualAccounts.length > 0
       },
-      ownCurrenciesPartners(): Array<any> {
-        let currencyIds = this.$store.getters.activeVirtualAccounts.map(
-          (a: any) => a.currencyId
-        )
-        return this.partners.filter((p: any) => {
-          return currencyIds.indexOf(p.backendId) > -1
-        })
-      },
-      ...mapGetters(["creditableMoneyAccounts"]),
       ...mapState(["recipientHistory"]),
     },
     methods: {
-      resetCredit(): void {
-        this.$store.state.showCredit = false
-        this.urlForHyperlink = ""
-        this.linkGenerated = false
-        this.amountForCredit = 0
-        this.errors.credit.amount = false
-        this.selectedCreditAccount =
-          this.creditableMoneyAccounts.length === 1
-            ? this.creditableMoneyAccounts[0]
-            : false
-      },
-
-      resetSendMoney(): void {
-        this.amount = 0
-        this.message = ""
-        this.searchName = ""
-        this.activeClass = 0
-        this.errors.transfer.balance = false
-        this.errors.transfer.amount = false
-      },
-
-      async newLinkTab() {
-        this.errors.credit.amount = false
-        if (this.amountForCredit <= 0) {
-          this.errors.credit.amount =
-            "Le montant à créditer doit être un nombre positif"
-          return
-        }
-        // This to ensure we are left with 2 decimals only
-        this.amountForCredit = this.amountForCredit.toFixed(2)
-        try {
-          if (!this.selectedCreditAccount) {
-            if (this.creditableMoneyAccounts.length > 1) {
-              throw new Error("Unexpected multiple creditable account found.")
-            }
-            this.selectedCreditAccount = this.creditableMoneyAccounts[0]
-          }
-          this.$loading.show()
-          let url = await this.selectedCreditAccount._obj.getCreditUrl(
-            this.amountForCredit
-          )
-          this.urlForHyperlink = url.order_url
-        } catch (error) {
-          console.log("Payment failed:", error)
-          this.$msg.error(
-            "Il y a eu un problème lors de la tentative de crédit de votre compte"
-          )
-        } finally {
-          this.$loading.hide()
-        }
-      },
-
-      navigateToCreditOrder(): void {
-        window.open(this.urlForHyperlink, "_blank")
-        this.urlForHyperlink = ""
-        this.amountForCredit = 0
-        this.showCreditRefreshNotification = true
-      },
-
-      closeAndRefresh(): void {
-        this.showCreditRefreshNotification = false
-        this.showModalFrameCreditMoney1 = false
-        this.$lokapi.flushBackendCaches()
-        this.$store.dispatch("fetchAccounts")
-        this.$store.dispatch("resetTransactions")
-        this.$store.commit("setModalState", false)
-      },
 
       copyUrl() {
         const el = document.createElement("textarea")
@@ -598,26 +139,6 @@
             })
         }
       },
-
-      async toggleFavorite(contact: any): Promise<boolean> {
-        try {
-          await contact.toggleFavorite()
-        } catch (e) {
-          let errorMessage = ""
-          if (!contact.jsonData.odoo.is_favorite)
-            errorMessage = "Une erreur est survenue lors de la mise en favoris,"
-          else
-            errorMessage =
-              "Une erreur est survenue lors de la suppression du favori,"
-          this.$msg.error(
-            errorMessage +
-              " veuillez ré-essayer ou contacter votre administrateur"
-          )
-          return false
-        }
-        return true
-      },
-
       async searchRecipientsHistory(): Promise<void> {
         let h = []
         for (let i = 0; i < this.recipientHistory.length; i++) {
@@ -633,126 +154,6 @@
           }
         }
         //this.partners = h
-      },
-
-      async searchRecipients(): Promise<void> {
-        this.partners = []
-        this.searchRecipientError = false
-        var recipients
-        try {
-          this.isLoading = true
-          recipients = await this.$lokapi.searchRecipients(this.searchName)
-        } catch (err) {
-          this.searchRecipientError = true
-          console.log("searchRecipients() Failed", err)
-        }
-        this.isLoading = false
-        this.partners = this.displayFavoritesOnly
-          ? returnFavoritesOnly(recipients)
-          : recipients
-      },
-
-      async handleClickRecipient(recipient: any): Promise<void> {
-        await this.setRecipient(recipient)
-        this.showModalFrame2 = true
-        this.showModalFrame1 = false
-        this.errors.transfer.balance = false
-        this.errors.transfer.amount = false
-        this.setFocusSend()
-      },
-
-      async setRecipient(partner: any): Promise<void> {
-        this.$store.state.lokapi.recipient = partner
-        this.recipientName = partner.name
-        this.recipientCurrencySymbol = await partner.getSymbol()
-      },
-
-      async sendTransaction(): Promise<void> {
-        this.errors.transfer.amount = false
-        this.errors.transfer.balance = false
-        if (this.amount <= 0) {
-          this.errors.transfer.amount =
-            "Le montant à transférer doit être un nombre positif"
-          return
-        }
-        // This to ensure we are left with 2 decimals only
-        this.amount = this.amount.toFixed(2)
-        let recipient = this.$store.state.lokapi.recipient
-        try {
-          this.$store.commit("setRequestLoadingAfterCreds", true)
-          await recipient.transfer(this.amount.toString(), this.message)
-        } catch (err) {
-          if (err instanceof LokapiExc.InsufficientBalance) {
-            this.errors.transfer.balance =
-              "Transaction refusée en raison de fonds insuffisants"
-            return
-          }
-          if (err.message === "User canceled the dialog box") {
-            // A warning message should have already been sent
-            return
-          }
-          this.$msg.error(
-            `Une erreur inattendue est survenue pendant le transfert d'argent. ` +
-              `Veuillez nous excuser pour la gêne occasionnée.<br>` +
-              `Vous pouvez réessayer. Si l'erreur persiste, veuillez contacter votre administrateur.`
-          )
-          console.log("Payment failed:", err.message)
-          throw err
-        } finally {
-          this.$loading.hide()
-          this.$store.commit("setRequestLoadingAfterCreds", false)
-        }
-        this.errors.transfer.balance = false
-        this.errors.transfer.amount = false
-        this.showModalFrame1 = false
-        this.showModalFrame2 = false
-
-        this.$msg.success(`Paiement effectué à ${this.recipientName}`)
-        if (!recipient.is_favorite) {
-          this.$Swal
-            .fire({
-              title: `Voulez vous ajouter ${this.recipientName} aux favoris ?`,
-              showDenyButton: true,
-              confirmButtonText: `Ajouter`,
-              denyButtonText: `Plus tard`,
-            })
-            .then(async (result: any) => {
-              if (!result.isConfirmed) {
-                if (await this.toggleFavorite(recipient)) {
-                  this.$Swal.fire(
-                    `${this.recipientName} a bien été ajouté en favori`,
-                    "",
-                    "success"
-                  )
-                }
-              }
-            })
-        }
-        await this.$store.dispatch("fetchAccounts")
-        await this.$store.dispatch("resetTransactions")
-        this.searchName = ""
-        this.partners = []
-        this.amount = 0
-        this.activeClass = 0
-      },
-      setSelectedCreditAccount(account: any): void {
-        this.selectedCreditAccount = account
-      },
-      setFocus() {
-        this.$nextTick(() => {
-          if (this.$refs.amountcredit) {
-            this.$refs.amountcredit.value = null
-            this.$refs.amountcredit.focus()
-            this.$refs.amountcredit.select()
-          }
-        })
-      },
-      setFocusSend() {
-        this.$nextTick(() => {
-          this.$refs.amountSend.value = null
-          this.$refs.amountSend.focus()
-          this.$refs.amountSend.select()
-        })
       },
     },
   })
@@ -770,38 +171,4 @@
     margin-left: 0
     margin-right: 0
 
-
-  div.account-selector
-    & :deep(.account)
-      min-width: fit-content
-      cursor: pointer
-
-    &.unselected :deep(.account)
-      opacity: 0.6
-      box-shadow: none
-      border: 2px #eee solid
-
-  .button.action
-    white-space: normal
-    height: auto
-  .card-recipient-wrapper
-    width: 90%
-  .favorit-icon-wrapper
-    width: 10%
-  .modal-card-body
-    min-height: 120px
-  .loader-container
-    position: relative
-    height: 80px
-  .amount-currency-symbol
-    margin: auto
-    font-size: 1.25em
-    font-weight: bold
-    line-height: 1em
-    padding-bottom: calc(0.5em - 1px)
-    padding-left: calc(0.75em - 1px)
-    padding-right: calc(0.75em - 1px)
-    padding-top: calc(0.5em - 1px)
-  .w-100
-    width: 100%
 </style>
